+++ packages/dialect-react-sdk/hooks/index.ts
@@ -1,5 +1,6 @@
 export { default as useDialectSdk } from './useDialectSdk';
 export { default as useDialectDapp } from './useDialectDapp';
+export { default as useIdentity } from './useIdentity';
 export { default as useThreads } from './useThreads';
 export { default as useThread } from './useThread';
 export { default as useThreadMessages } from './useThreadMessages';