+++ packages/dialect-react-ui/components/common/index.ts
@@ -5,4 +5,3 @@ export { default as LinkifiedText } from './LinkifiedText';
 export { default as ToggleSection } from './ToggleSection';
 export { default as Section } from './Section';
 export { default as OnChainIndicator } from './OnChainIndicator';
-export { fetchSolanaNameServiceName } from './SolanaNameServiceName';